[FLINK-8989] [e2eTests] Elasticsearch1&2&5 end to end test
